fix: add is_live in report data request
