+++ src/services/api/resources/dashboards.js
@@ -3,6 +3,7 @@ import http from '@/services/api/http';
 import Config from '@/store/modules/config';
 import DashboardStore from '@/store/modules/dashboards';
 
+import { isFilteringDates } from '@/utils/filter';
 import { createRequestQuery } from '@/utils/request';
 
 export default {
@@ -102,18 +103,10 @@ export default {
     const { appliedFilters } = DashboardStore.state;
     const { currentDashboardFilters } = DashboardStore.state;
 
-    const currentDashboardDateFiltersNames = currentDashboardFilters
-      .filter((filter) => filter.type === 'date_range')
-      .map((filter) => filter.name);
-
-    const appliedFilterNames = Object.keys(appliedFilters);
-
-    const isFilteringDates = currentDashboardDateFiltersNames.some(
-      (currentDashboardDateFilterName) =>
-        appliedFilterNames.some((appliedFilterName) =>
-          currentDashboardDateFilterName.includes(appliedFilterName),
-        ),
-    );
+    const hasDateFilter = isFilteringDates({
+      currentDashboardFilters,
+      appliedFilters,
+    });
 
     if (!dashboardUuid || !widgetUuid) {
       throw new Error(
@@ -123,7 +116,7 @@ export default {
 
     const treatedParams = createRequestQuery(appliedFilters, {
       project: Config.state.project.uuid,
-      is_live: !isFilteringDates || undefined,
+      is_live: !hasDateFilter || undefined,
       ...params,
     });
     const widgetData = await http.get(
@@ -162,6 +155,13 @@ export default {
     next,
   }) {
     const { appliedFilters } = DashboardStore.state;
+    const { currentDashboardFilters } = DashboardStore.state;
+
+    const hasDateFilter = isFilteringDates({
+      currentDashboardFilters,
+      appliedFilters,
+    });
+
     if (!dashboardUuid || !widgetUuid) {
       throw new Error(
         'Please provide valids UUIDs parameters to request report data of widget.',
@@ -174,6 +174,7 @@ export default {
       offset,
       limit,
       next,
+      is_live: !hasDateFilter || undefined,
     });
 
     const reportData = await http.get(
